rubik: add psi indicator observation framework
We built a Pod adjustment framework based on indicators-triggers, and built a PSI service framework based on this.
Triggers are managed through a tree structure and invoked through a chain of responsibility pattern.

First, we implement a Metric interface for updating PSI indicator data. For exceeding the standard, we trigger follow-up operations through different Triggers, including expulsion of nodes, statistics of resources, etc.

Also, we removed redundant code and fixed syntax errors.

This time does not implement resource statistics and eviction trigger interfaces.

Signed-off-by: vegbir <yangjiaqi16@huawei.com>